requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-dotenv
flask-migrate
flask-sqlalchemy
requests==2.31.0
openai
pdfplumber
python-docx
